Don't check for compliance with BIP69's lexicographical ordering
I was convinced that BIP69 is not such a great idea (primarily due to
concerns relating to the transition period) and that there might be
better alternatives to it (like deterministically sorting based on a
user secret). At the very least, it seems like BIP69's usefulness is
controversial enough to not warrant a warning message for non-compliance.

Also see: https://github.com/bitcoin/bitcoin/issues/12457