Revert "Support direct nojs redirect with external HTML rewriting"
It does not work well with the testnet/liquid subdirectories setup, which nginx
includes in $request_uri, which prevents the redirect from working properly.
When taking this into account, there doesn't seem to be an easy way to get nginx
to do the HTML rewriting that doesn't involve Lua scripting, which makes this
probably not worth it.

This reverts commit a925c7980ae538527948b4ed10985b0319c2c693.